refactor: rename range-datepicker to date-range-picker
